fix(frontend): resolve CI failures and optimize bundle size
 - Fixed performance budget check in performance.yml\n - Corrected useReportWebVitals import in PerformanceMonitor.tsx\n - Optimized layout with dynamic imports for Toaster and TransactionHistoryDrawer\n - Split TransactionHistoryDrawer to reduce main bundle size\n - Fixed ESLint 9 flat config compatibility\n - Fixed typo in Home page
